Return httpStatus code with error
